+++ Kernel/src/simulationTools/Lsodar.h
@@ -43,18 +43,18 @@ public:
 
   /** \fn double getR()
   *   \brief Return the r of the OneStepIntegrator
-  *   \return double : the value of r
+  *   \return int : the value of r
   */
-  inline double getR(void) const
+  inline int getR(void) const
   {
     return this->r;
   }
 
-  /** \fn void setR(double r)
+  /** \fn void setR(int r)
   *   \brief Return the r of OneStepIntegrator
   *   \param double : the value to set r
   */
-  inline void setR(const double r)
+  inline void setR(const int r)
   {
     this->r = r;
   }
